// apps/pei-collab/src/modules/plano-aee/types/materials.types.ts
export type MaterialType = 
  | 'visual' 
  | 'tactile' 
  | 'audio' 
  | 'digital' 
  | 'adaptation' 
  | 'other';

export type MaterialProductionStatus = 
  | 'planned' 
  | 'in_progress' 
  | 'completed' 
  | 'cancelled';